Move helpers/ to src/ & update references
