+++ client/zestyreactnative/src/helpers/scaffolding.js
